automation-portfolio-file-organizer/requirements.txt
# Core dependencies
pyyaml>=6.0
click>=8.0
pathlib2>=2.3.0; python_version<"3.4"

# Development dependencies
pytest>=7.0
pytest-cov>=4.0
black>=22.0
flake8>=5.0
mypy>=1.0

# Documentation dependencies
sphinx>=5.0
sphinx-rtd-theme>=1.0

# Optional dependencies
watchdog>=2.0  # For file system monitoring
pillow>=9.0  # For image metadata extraction
python-magic>=0.4  # For file type detection